Reorganize encoding type traits
