+++ crates/any/any-counterparty/src/impls/encoding/client_state.rs
@@ -3,8 +3,8 @@ use core::marker::PhantomData;
 use cgp_core::error::CanRaiseError;
 use hermes_encoding_components::traits::convert::Converter;
 use hermes_encoding_components::traits::decoder::{CanDecode, Decoder};
-use hermes_encoding_components::traits::encoded::HasEncodedType;
 use hermes_encoding_components::traits::schema::HasSchema;
+use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use hermes_protobuf_encoding_components::types::{Any, ViaProtobuf};
 use hermes_protobuf_encoding_components::vendor::HasSchemaType;
 use ibc_relayer_types::clients::ics07_tendermint::client_state::ClientState as TendermintClientState;
