+++ crates/cosmos/cosmos-chain-components/src/encoding/components.rs
@@ -4,9 +4,10 @@ use hermes_encoding_components::impls::encoded::ProvideEncodedBytes;
 use hermes_encoding_components::impls::schema::ProvideStringSchema;
 pub use hermes_encoding_components::traits::convert::ConverterComponent;
 pub use hermes_encoding_components::traits::decoder::DecoderComponent;
-pub use hermes_encoding_components::traits::encoded::EncodedTypeComponent;
 pub use hermes_encoding_components::traits::encoder::EncoderComponent;
-pub use hermes_encoding_components::traits::schema::{SchemaGetterComponent, SchemaTypeComponent};
+pub use hermes_encoding_components::traits::schema::SchemaGetterComponent;
+pub use hermes_encoding_components::traits::types::encoded::EncodedTypeComponent;
+pub use hermes_encoding_components::traits::types::schema::SchemaTypeComponent;
 
 use crate::encoding::convert::CosmosConverterComponents;
 use crate::encoding::encode::CosmosEncoderComponents;
