+++ crates/cosmos/cosmos-chain-components/src/impls/queries/abci.rs
@@ -1,7 +1,7 @@
 use cgp_core::error::CanRaiseError;
-use hermes_encoding_components::traits::encoded::HasEncodedType;
 use hermes_encoding_components::traits::encoder::CanEncode;
 use hermes_encoding_components::traits::has_encoding::HasEncoding;
+use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use hermes_encoding_components::types::AsBytes;
 use hermes_protobuf_encoding_components::types::ViaProtobuf;
 use hermes_relayer_components::chain::traits::types::height::HasHeightType;
