+++ crates/cosmos/cosmos-wasm-relayer/src/impls/create_client_message.rs
@@ -6,9 +6,9 @@ use hermes_cosmos_chain_components::types::tendermint::{
     TendermintClientState, TendermintConsensusState,
 };
 use hermes_encoding_components::traits::convert::CanConvert;
-use hermes_encoding_components::traits::encoded::HasEncodedType;
 use hermes_encoding_components::traits::encoder::CanEncode;
 use hermes_encoding_components::traits::has_encoding::HasDefaultEncoding;
+use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use hermes_encoding_components::types::AsBytes;
 use hermes_protobuf_encoding_components::types::ViaAny;
 use hermes_relayer_components::chain::traits::message_builders::create_client::CreateClientMessageBuilder;
