+++ crates/encoding/encoding-components/src/impls/delegate.rs
@@ -4,9 +4,10 @@ use cgp_core::prelude::{DelegateComponent, HasErrorType};
 
 use crate::traits::convert::Converter;
 use crate::traits::decoder::Decoder;
-use crate::traits::encoded::HasEncodedType;
 use crate::traits::encoder::Encoder;
-use crate::traits::schema::{HasSchemaType, SchemaGetter};
+use crate::traits::schema::SchemaGetter;
+use crate::traits::types::encoded::HasEncodedType;
+use crate::traits::types::schema::HasSchemaType;
 
 pub struct DelegateEncoding<Delegate>(pub PhantomData<Delegate>);
