+++ crates/encoding/encoding-components/src/impls/return_encoded.rs
@@ -1,8 +1,8 @@
 use cgp_core::error::HasErrorType;
 
 use crate::traits::decoder::Decoder;
-use crate::traits::encoded::HasEncodedType;
 use crate::traits::encoder::Encoder;
+use crate::traits::types::encoded::HasEncodedType;
 
 pub struct ReturnEncoded;
