+++ crates/encoding/encoding-components/src/traits/encode_and_decode_mut.rs
@@ -1,7 +1,9 @@
 use cgp_core::error::HasErrorType;
 
-use crate::traits::decode_mut::{CanDecodeMut, HasDecodeBufferType, MutDecoder};
-use crate::traits::encode_mut::{CanEncodeMut, HasEncodeBufferType, MutEncoder};
+use crate::traits::decode_mut::{CanDecodeMut, MutDecoder};
+use crate::traits::encode_mut::{CanEncodeMut, MutEncoder};
+use crate::traits::types::decode_buffer::HasDecodeBufferType;
+use crate::traits::types::encode_buffer::HasEncodeBufferType;
 
 pub trait CanEncodeAndDecodeMut<Strategy, Value>:
     CanEncodeMut<Strategy, Value> + CanDecodeMut<Strategy, Value>
