+++ crates/encoding/encoding-components/src/traits/encoder.rs
@@ -1,6 +1,6 @@
 use cgp_core::prelude::*;
 
-use crate::traits::encoded::HasEncodedType;
+use crate::traits::types::encoded::HasEncodedType;
 
 #[derive_component(EncoderComponent, Encoder<Encoding>)]
 pub trait CanEncode<Strategy, Value>: HasEncodedType + HasErrorType {
