+++ crates/encoding/encoding-components/src/traits/mod.rs
@@ -4,7 +4,7 @@ pub mod decoder;
 pub mod encode_and_decode;
 pub mod encode_and_decode_mut;
 pub mod encode_mut;
-pub mod encoded;
 pub mod encoder;
 pub mod has_encoding;
 pub mod schema;
+pub mod types;
