+++ crates/encoding/encoding-components/src/traits/schema.rs
@@ -2,10 +2,7 @@ use core::marker::PhantomData;
 
 use cgp_core::prelude::*;
 
-#[derive_component(SchemaTypeComponent, ProvideSchemaType<Encoding>)]
-pub trait HasSchemaType: Async {
-    type Schema: Async;
-}
+use crate::traits::types::schema::HasSchemaType;
 
 #[derive_component(SchemaGetterComponent, SchemaGetter<Encoding>)]
 pub trait HasSchema<Value>: HasSchemaType {
