+++ crates/encoding/encoding-components/src/traits/types/encoded.rs
