+++ crates/encoding/encoding-components/src/traits/types/schema.rs
@@ -1,0 +1,6 @@
+use cgp_core::prelude::*;
+
+#[derive_component(SchemaTypeComponent, ProvideSchemaType<Encoding>)]
+pub trait HasSchemaType: Async {
+    type Schema: Async;
+}
