+++ crates/encoding/protobuf-encoding-components/src/impls/protobuf.rs
@@ -1,7 +1,7 @@
 use cgp_core::error::{CanRaiseError, HasErrorType};
 use hermes_encoding_components::traits::decoder::Decoder;
-use hermes_encoding_components::traits::encoded::HasEncodedType;
 use hermes_encoding_components::traits::encoder::Encoder;
+use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use prost::{DecodeError, Message};
 
 pub struct EncodeAsProtobuf;
