+++ crates/encoding/protobuf-encoding-components/src/vendor.rs
@@ -1,1 +1,2 @@
-pub use hermes_encoding_components::traits::schema::{HasSchemaType, SchemaGetter};
+pub use hermes_encoding_components::traits::schema::SchemaGetter;
+pub use hermes_encoding_components::traits::types::schema::HasSchemaType;
