+++ crates/relayer/relayer-components/src/chain/impls/queries/query_and_decode_consensus_state.rs
@@ -4,8 +4,8 @@ use hermes_encoding_components::types::AsBytes;
 
 use cgp_core::prelude::{Async, CanRaiseError};
 use hermes_encoding_components::traits::decoder::CanDecode;
-use hermes_encoding_components::traits::encoded::HasEncodedType;
 use hermes_encoding_components::traits::has_encoding::HasDefaultEncoding;
+use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 
 use crate::chain::traits::queries::consensus_state::{
     CanQueryRawConsensusState, ConsensusStateQuerier,
