+++ crates/wasm/wasm-client-components/src/types/consensus_state.rs
@@ -1,8 +1,8 @@
 use cgp_core::error::HasErrorType;
 use hermes_encoding_components::traits::convert::{CanConvert, Converter};
 use hermes_encoding_components::traits::decoder::{CanDecode, Decoder};
-use hermes_encoding_components::traits::encoded::HasEncodedType;
 use hermes_encoding_components::traits::encoder::{CanEncode, Encoder};
+use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use hermes_protobuf_encoding_components::types::{Any, ViaAny};
 
 #[allow(clippy::derive_partial_eq_without_eq)]
